[CARDS-603:refactor] move non-context code out of the questionnaire context files
QuestionnaireTreeContext.jsx and QuestionnaireContext.jsx held a lot of code
that is not really React context: pure tree helpers, the server API calls,
display config, and a scroll-tracking hook. This made both files hard to read.

Move that code into focused modules and keep the context files for context
only. No behavior change - this just moves code and updates imports.

New modules in questionnaireEditor/:
- questionnaireApi.js       - jcrActions server calls
- entryDisplay.jsx          - ENTRY_TITLE_FIELD_SPEC + jcrGetConditionalTitle
- treeQueries.js            - isDescendant, findNode*, findTreeEntries,
                                getEntryChildIds, getAncestorPath
- questionnaireTreeModel.js - JCR-to-node building, subtree splicing, validators

New module in questionnaire/:
- useInViewTracker.js       - the scroll / in-view / highlight hook

Also:
- move getOrdinalString into QuestionnaireUtilities.jsx
- delete the unused useEntryStyles hook
- rename the test file to treeQueries.test.js to match what it now tests

QuestionnaireTreeContext.jsx: 937 -> 302 lines
QuestionnaireContext.jsx:     198 -> 85 lines
